Replace ArrayList Creation on a return
Array list was being returned using non standard way of creating the
array as essentially an internal class.  No real good reason for this
and it was being flagged via sonar for not exposing private constructor
to prevent synthetic access.  Much cleaner approach is to just build the
array without the internal class.